feat: Z-wave integration
In order to use z-wave devices in Home Assistant, our setup needs two
changes:
1. The USB Z-Wave hub has to be made available to Nomad
2. We need a Z-Wave server to handle communications between the hub and
   Home Assistant.

The USB device can only be passed to one machine/VM, so we add a tag to
the Nomad client that has access to it. The Z-Wave server is then
constrained to only run on that machine. We can then use our normal
service discovery to integrate into Home Assistant.