feat: updates required to rebuild cluster
Updates required to rebuild the cluster and provide more storage space
on the cluster instances.

The main challenge was allowing the VMs to have more storage space than
the template used to create them. This requires the main partition to be
last (after the boot and swap partitions) so that it can be expanded by
cloud-init after the first boot.